@@ -6,10 +6,11 @@
 	"directories": {
 		"test": "tests"
 	},
+	"type": "module",
 	"scripts": {
-		"dev": "NODE_ENV=development webpack --progress --config webpack.js",
-		"watch": "NODE_ENV=development webpack --progress --watch --config webpack.js",
-		"build": "NODE_ENV=production webpack --progress --config webpack.js",
+		"build": "NODE_ENV=production vite --mode production build",
+		"dev": "NODE_ENV=development vite --mode development build",
+		"watch": "NODE_ENV=development vite --mode development build --watch",
 		"lint": "eslint --ext .js,.vue src",
 		"lint:fix": "eslint --ext .js,.vue src --fix",
 		"stylelint": "stylelint src/**/*.vue src/**/*.scss src/**/*.css",
@@ -40,17 +41,17 @@
 		"@maplibre/maplibre-gl-geocoder": "^1.8.0",
 		"@nextcloud/auth": "^2.5.1",
 		"@nextcloud/axios": "^2.1.0",
-		"@nextcloud/dialogs": "^6.3.0",
+		"@nextcloud/dialogs": "^7.0.0-rc.1",
 		"@nextcloud/event-bus": "^3.3.2",
 		"@nextcloud/initial-state": "^2.2.0",
 		"@nextcloud/l10n": "^3.1.0",
 		"@nextcloud/moment": "^1.2.1",
 		"@nextcloud/paths": "^2.2.1",
 		"@nextcloud/router": "^3.0.0",
-		"@nextcloud/vue": "^8.27.0",
+		"@nextcloud/vue": "^9.0.0-rc.2",
 		"@raruto/leaflet-elevation": "1.5.3",
 		"babel-eslint": "^10.1.0",
-		"chart.js": "^3.9.1",
+		"chart.js": "^4.5.0",
 		"ds-countdown": "^1.1.1",
 		"jquery": "^3.5.1",
 		"kjua": "^0.10.0",
@@ -68,8 +69,8 @@
 		"mapbox-gl": "2.4.1",
 		"mapbox-gl-leaflet": "^0.0.15",
 		"maplibre-gl": "^5.6.0",
-		"vue": "^2.7.16",
-		"vue-chartjs": "^4.1.2",
+		"vue": "^3.5.17",
+		"vue-chartjs": "^5.3.2",
 		"vue-clipboard2": "^0.3.3",
 		"vue-material-design-icons": "^5.3.1"
 	},
@@ -78,7 +79,7 @@
 		"@nextcloud/browserslist-config": "^3.0.0",
 		"@nextcloud/eslint-config": "^8.0.0",
 		"@nextcloud/stylelint-config": "^3.0.1",
-		"@nextcloud/vite-config": "^1.4.2",
+		"@nextcloud/vite-config": "^2.3.5",
 		"@nextcloud/webpack-vue-config": "^6.2.0",
 		"eslint-webpack-plugin": "^4.2.0",
 		"stylelint-webpack-plugin": "^5.0.1",
