@@ -182,7 +182,7 @@ export default {
 			this.styles = {
 				...getVectorStyles(apiKey, !this.isPublicPage && this.settings.proxy_osm),
 				...getRasterTileServers(apiKey, !this.isPublicPage && this.settings.proxy_osm),
-				...getExtraTileServers(this.settings.extra_tile_servers, apiKey, !this.isPublicPage && this.settings.proxy_osm),
+				...getExtraTileServers(this.settings.extra_tile_servers ?? [], apiKey, !this.isPublicPage && this.settings.proxy_osm),
 			}
 			const restoredStyleKey = Object.keys(this.styles).includes(this.settings.mapStyle) ? this.settings.mapStyle : 'streets'
 			const restoredStyleObj = this.styles[restoredStyleKey]
